requirements.txt
PyYAML~=6.0.1
aiogram~=3.2.0
openai~=1.6.1
langchain~=0.0.352
